data_pipelines/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
starlette<=0.33.0,>=0.29.0
attrs==23.1.0
cattrs==23.2.3
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.6.1
web_pdb==1.6.2
pandas==1.5.3
selenium==4.21.0
beautifulsoup4==4.12.3
